+++ Sources/Sentry/include/SentryCrashWrapper.h
@@ -25,11 +25,7 @@ SENTRY_NO_INIT
 
 - (void)installAsyncHooks;
 
-/**
- * It's not really possible to close SentryCrash. Best we can do is to deactivate all the monitors,
- * clear the `onCrash` callback installed on the global handler, and a few more minor things.
- */
-- (void)close;
+- (void)uninstallAsyncHooks;
 
 - (NSDictionary *)systemInfo;