+++ Sources/SentryCrash/Recording/SentryCrash.h
@@ -43,6 +43,8 @@ typedef enum {
 
 static NSString *const SENTRYCRASH_REPORT_ATTACHMENTS_ITEM = @"attachments";
 
+@class SentryNSNotificationCenterWrapper;
+
 /**
  * Reports any crashes that occur in the application.
  *
@@ -139,10 +141,6 @@ static NSString *const SENTRYCRASH_REPORT_ATTACHMENTS_ITEM = @"attachments";
  */
 @property (nonatomic, readwrite, assign) SentryCrashReportWriteCallback onCrash;
 
-/** Add a copy of SentryCrash's console log messages to the crash report.
- */
-@property (nonatomic, readwrite, assign) BOOL addConsoleLogToReport;
-
 /** Print the previous app run log to the console when installing SentryCrash.
  *  This is primarily for debugging purposes.
  */
@@ -207,6 +205,15 @@ static NSString *const SENTRYCRASH_REPORT_ATTACHMENTS_ITEM = @"attachments";
  */
 - (BOOL)install;
 
+/**
+ * It's not really possible to completely uninstall SentryCrash. The best we can do is to deactivate
+ * all the monitors, keep track of the already installed monitors to install them again if someone
+ * calls install, clear the `onCrash` callback, and stop the SentryCrashCachedData. For the
+ * ``SentryCrashMonitorTypeMachException`` we let the exception threads running, but they don't
+ * report anything.
+ */
+- (void)uninstall;
+
 /** Send all outstanding crash reports to the current sink.
  * It will only attempt to send the most recent 5 reports. All others will be
  * deleted. Once the reports are successfully sent to the server, they may be
@@ -243,6 +250,12 @@ static NSString *const SENTRYCRASH_REPORT_ATTACHMENTS_ITEM = @"attachments";
  * @param reportID An ID of report to delete.
  */
 - (void)deleteReportWithID:(NSNumber *)reportID;
+
+/**
+ * Only needed for testing.
+ */
+- (void)setSentryNSNotificationCenterWrapper:
+    (SentryNSNotificationCenterWrapper *)notificationCenter;
 
 @end